requirements.txt
streamlit
openai
pandas
umap-learn
plotly
transformers
torch
datasets
